# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Chowdhary Mart API — hyperlocal e-commerce platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      operationId: updateMe
      tags: [auth]
      summary: Update profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # ── Categories ────────────────────────────────────────────────────────────
...
  /stores:
    get:
      operationId: listStores
      tags: [stores]
      summary: List nearby stores
      parameters:
        - name: lat
          in: query
          schema:
            type: number
        - name: lng
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Stores list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"
...
  /addresses:
    get:
      operationId: listAddresses
      tags: [addresses]
      summary: List my addresses
      responses:
        "200":
          description: Addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Address"
    post:
      operationId: createAddress
      tags: [addresses]
      summary: Add an address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
  /addresses/{addressId}:
    patch:
      operationId: updateAddress
      tags: [addresses]
      summary: Update address
      parameters:
        - $ref: "#/components/parameters/AddressIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
    delete:
      operationId: deleteAddress
      tags: [addresses]
      summary: Delete address
      parameters:
        - $ref: "#/components/parameters/AddressIdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Cart ──────────────────────────────────────────────────────────────────
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List my orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: placeOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
  /admin/orders:
    get:
      operationId: listAdminOrders
      tags: [admin]
      summary: List all orders (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
...
  /admin/coupons:
    get:
      operationId: listAdminCoupons
      tags: [admin]
      summary: List all coupons (admin)
      responses:
        "200":
          description: Coupons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
    post:
      operationId: createCoupon
      tags: [admin]
      summary: Create coupon (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
...
components:
  parameters:
...
    AddressIdParam:
      name: addressId
      in: path
      required: true
      schema:
        type: integer
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── Auth ───────────────────────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [name, password]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
        phone:
          type: string
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [customer, vendor, delivery_partner]
    LoginInput:
      type: object
      required: [password]
      properties:
        email:
          type: string
        phone:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
    UserProfile:
      type: object
      required: [id, name, role]
      properties:
        id:
          type: integer
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
        walletBalance:
          type: string
        loyaltyPoints:
          type: integer
        referralCode:
          type: ["string", "null"]
        isVerified:
          type: boolean
        createdAt:
          type: string
    UpdateProfileInput:
      type: object
      properties:
        name:
          type: string
        avatarUrl:
          type: string
        phone:
          type: string

    # ── Category ───────────────────────────────────────────────────────────
...
    Store:
      type: object
      required: [id, name, lat, lng, address]
      properties:
        id:
          type: integer
        name:
          type: string
        logoUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        address:
          type: string
        city:
          type: ["string", "null"]
        rating:
          type: ["string", "null"]
        ratingCount:
          type: ["integer", "null"]
        isOpen:
          type: boolean
        estimatedDeliveryMins:
          type: ["integer", "null"]
        deliveryFee:
          type: ["string", "null"]
        freeDeliveryAbove:
          type: ["string", "null"]
        minOrderValue:
          type: ["string", "null"]
        lat:
          type: number
        lng:
          type: number
...
    Address:
      type: object
      required: [id, label, name, line1, city, state, pincode]
      properties:
        id:
          type: integer
        label:
          type: string
        name:
          type: string
        phone:
          type: ["string", "null"]
        line1:
          type: string
        line2:
          type: ["string", "null"]
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        isDefault:
          type: boolean
    AddressInput:
      type: object
      required: [label, name, line1, city, state, pincode]
      properties:
        label:
          type: string
        name:
          type: string
        phone:
          type: string
        line1:
          type: string
        line2:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        lat:
          type: number
        lng:
          type: number
        isDefault:
          type: boolean

    # ── Cart ───────────────────────────────────────────────────────────────
...
    Order:
      type: object
      required: [id, orderNumber, status, total, createdAt]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        userId:
          type: integer
        storeId:
          type: integer
        status:
          type: string
        paymentMethod:
          type: string
        paymentStatus:
          type: string
        subtotal:
          type: string
        deliveryFee:
          type: string
        discount:
          type: string
        couponCode:
          type: ["string", "null"]
        couponDiscount:
          type: ["string", "null"]
        walletUsed:
          type: ["string", "null"]
        total:
          type: string
        loyaltyPointsEarned:
          type: ["integer", "null"]
        estimatedDeliveryMins:
          type: ["integer", "null"]
        createdAt:
          type: string
        store:
          $ref: "#/components/schemas/Store"
...
    OrderInput:
      type: object
      required: [addressId, paymentMethod]
      properties:
        addressId:
          type: integer
        paymentMethod:
          type: string
          enum: [cod, online, wallet, upi]
        couponCode:
          type: string
        useWallet:
          type: boolean
        notes:
          type: string
...
    Coupon:
      type: object
      required: [id, code, discountType, discountValue]
      properties:
        id:
          type: integer
        code:
          type: string
        description:
          type: ["string", "null"]
        discountType:
          type: string
        discountValue:
          type: string
        minOrderValue:
          type: ["string", "null"]
        maxDiscount:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
    CouponInput:
      type: object
      required: [code, discountType, discountValue]
      properties:
        code:
          type: string
        description:
          type: string
        discountType:
          type: string
          enum: [percent, flat]
        discountValue:
          type: string
        minOrderValue:
          type: string
        maxDiscount:
          type: string
        usageLimit:
          type: integer
        perUserLimit:
          type: integer
        expiresAt:
          type: string
...